requirements.txt
absl-py==2.0.0
accelerate==0.23.0
aiohttp==3.8.5
aiosignal==1.3.1
anyascii==0.3.2
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.0
Babel==2.12.1
bangla==0.0.2
blinker==1.6.2
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.6
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
clean-fid==0.1.35
click==8.1.7
clip-anytorch==2.5.2
cmake==3.27.5
contourpy==1.1.1
coqpit==0.0.17
cycler==0.11.0
Cython==0.29.30
dateparser==1.1.8
dctorch==0.1.2
decorator==4.4.2
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.6.1
encodec==0.1.1
filelock==3.12.4
Flask==2.3.3
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
g2pkk==0.1.2
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.1
google-auth-oauthlib==1.0.0
gpt4all==1.0.12
grpcio==1.58.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
huggingface-hub==0.17.3
idna==3.4
imageio==2.31.4
imageio-ffmpeg==0.4.9
inflect==5.6.2
itsdangerous==2.1.2
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
jsonlines==1.2.0
jsonmerge==1.9.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
k-diffusion==0.1.0
kiwisolver==1.4.5
kornia==0.7.0
lazy_loader==0.3
librosa==0.10.1
lit==17.0.1
llvmlite==0.40.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.3
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.6
multidict==6.0.4
networkx==2.8.8
nltk==3.8.1
num2words==0.5.12
numba==0.57.0
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
packaging==23.1
pandas==1.4.4
pathtools==0.1.2
Pillow==10.0.1
platformdirs==3.10.0
pooch==1.7.0
proglog==0.1.10
protobuf==4.24.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pynndescent==0.5.10
pyparsing==3.1.1
pypinyin==0.49.0
pysbd==0.3.4
python-crfsuite==0.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rotary-embedding-torch==0.3.0
rpds-py==0.10.3
rsa==4.9
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
sentry-sdk==1.31.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
soxr==0.3.6
sympy==1.12
tbb==2021.10.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
tifffile==2023.9.26
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchdiffeq==0.2.3
torchsde==0.2.6
torchvision==0.15.2
tqdm==4.64.1
trainer==0.0.31
trampoline==0.1.2
transformers==4.33.2
triton==2.0.0
TTS==0.17.5
typing_extensions==4.8.0
tzlocal==5.0.1
umap-learn==0.5.4
Unidecode==1.3.7
urllib3==2.0.5
wandb==0.15.11
wcwidth==0.2.6
Werkzeug==2.3.7
yarl==1.9.2
